Improving code based on PR feedback. Mainly rename refactoring.
